remove simtk imports

+++ bgmol/systems/base.py
@@ -106,7 +106,7 @@ class OpenMMSystem(BaseSystem):
         self._system = mm.System()
 
         # Store positions.
-        self._positions = unit.Quantity(np.zeros([0, 3], np.float), unit.nanometers)
+        self._positions = unit.Quantity(np.zeros([0, 3], float), unit.nanometers)
 
         # Empty topology.
         self._topology = app.Topology()
